Fix by-sierra-location to use collectionType
This fixes an issue where this by-sierra-location factory was exposing
`collectionType`, but using the data found in `locationType` (a
deprecated nypl-core location property). In the absense of that
data, our factory was falling back on a sensible default of
['Branch']. This PR wires up the factory to correctly use the actual
data in the nypl-core `collectionType` property so that integrators can
act on whether or not a location's collection has "Research" materials.
Also adds some tests.

Also adds updated locations.json fixture.